Fix ngtcp2 and libheif build missing lib bug
